% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{file.path}
\alias{file.path}
\alias{\%//\%}
\title{Construct Path to File}
\usage{
x \%//\% y
}
\arguments{
\item{x, y}{Character vectors.}
}
\description{
Analogous to \code{file.path(x, y)}.
}
\examples{
"home" \%//\% "dir" # returns "home/dir"
}
