\name{infix-package}
\alias{infix-package}
\alias{infix}
\docType{package}
\title{ Set of infix operators to simplify the programming code. }
\description{ Provides infix operators to make coding easier. }
\details{
\tabular{ll}{
Package: \tab infix\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-01-10\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Maintainer: Luis Delgado
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
