% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix-package.R
\docType{package}
\name{infix}
\alias{infix}
\alias{infix-package}
\title{Basic Infix Binary Operators}
\description{
Contains a number of infix binary operators that may be useful
    in day to day practices.
}
\section{Operators}{
 \describe{
    \item{\code{\link{except}}}{Simple Error Handling.}
    \item{\code{\link{operators}}}{Common Infix Operators.}
    \item{\code{\link{funlogic}}}{Function Logical Operators.}
    \item{\code{\link{pipes}}}{Package's 'magrittr' pipe-operators.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ebeneditos/infix}
  \item Report bugs at \url{http://github.com/ebeneditos/infix/issues}
}

}
\author{
\strong{Maintainer}: Ernest Benedito \email{ebeneditos@gmail.com}

}
