% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{nil}
\alias{nil}
\alias{\%||\%}
\title{Default value for \code{NULL}}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{If \code{a} is \code{NULL}, will return \code{b}; otherwise returns \code{a}.}
}
\description{
This infix function makes it easy to replace \code{NULL}`s with
    a default value. It's inspired by the way that Ruby's or operation
    (\code{`||`}) works.
}
\examples{
1 \%||\% 2 # returns 1
NULL \%||\% 2 # returns 2
}
