% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{nomatch}
\alias{nomatch}
\alias{\%!in\%}
\title{Value (non) Matching}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{Vector or \code{NULL}: the values to be (not) matched.}

\item{table}{Vector or \code{NULL}: the values to be (not) matched against.}
}
\description{
The logical negation of \code{\link[base]{\%in\%}}.
}
\examples{
4 \%!in\% 1:3 # returns TRUE
}
