% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operators}
\alias{operators}
\title{Common Infix Operators}
\description{
Common Infix Operators
}
\section{Infix Operators}{
 \describe{
    \item{\code{\link{paste0}}}{Concatenate Strings (see \code{?"\%+\%"}).}
    \item{\code{\link{file.path}}}{Construct Path to File (see \code{?"\%//\%"}).}
    \item{\code{\link{nomatch}}}{Value (non) Matching (see \code{?"\%!in\%"}).}
    \item{\code{\link{nil}}}{Default value for \code{NULL} (see \code{?"\%||\%"}).}
}
}

