% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{paste0}
\alias{paste0}
\alias{\%+\%}
\alias{paste}
\title{Concatenate Strings}
\usage{
x \%+\% y
}
\arguments{
\item{x, y}{Objects to be converted to character vectors.}
}
\description{
Analogous to \code{paste0(x, y)}.
}
\examples{
"01" \%+\% "jan" \%+\% "1970" # returns "01jan1970"
}
