% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\docType{import}
\name{pipes}
\alias{pipes}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%T>\%}
\title{Package's 'magrittr' pipe-operators}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs

lhs \%$\% rhs

lhs \%T>\% rhs
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%T>\%}}}
}}

